fully functional sliders
When a player joins the game, a slider is displayed for that player,
designated with their name underneath it. Their designated slider (on
the display page) changes accurately with changes to their personal
slider (on their own controller page). The display page supports up to
6 players (exactly like the pong game) and handles disconnected players
accurately, allowing for a new player to take their spot.